+++ Routes/Items/GetItems.cs
@@ -1,10 +1,12 @@
 
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using Assets.Scripts.Objects;
+using Ceen;
 using WebAPI.Payloads;
 
-namespace WebAPI.Routes.Devices
+namespace WebAPI.Routes.Items
 {
     class GetItems : IWebRoute
     {
@@ -12,10 +14,10 @@ namespace WebAPI.Routes.Devices
 
         public string[] Segments => new[] { "items" };
 
-        public void OnRequested(RequestEventArgs e, IDictionary<string, string> pathParams)
+        public async Task OnRequested(IHttpContext context, IDictionary<string, string> pathParams)
         {
-            var items = Item.AllItems.Select(x => ItemPayload.FromItem(x));
-            e.Context.SendResponse(200, items);
+            var payload = await Dispatcher.RunOnMainThread(() => Item.AllItems.Select(x => ItemPayload.FromItem(x)));
+            await context.SendResponse(HttpStatusCode.OK, payload);
         }
     }
 }