+++ Payloads/PostSavePayload.cs
@@ -1,0 +1,8 @@
+
+namespace WebAPI.Payloads
+{
+    public class PostSavePayload
+    {
+        public string fileName { get; set; }
+    }
+}