+++ WebAPI.cs
@@ -1,12 +1,10 @@
 
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using BepInEx;
 using HarmonyLib;
-using UnityEngine;
 using WebAPI.Server.Attributes;
 using WebAPI.Server;
 using Assets.Scripts.Networking;