+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### vNext
+
+- Switch to Ceen http library for future ssl security.
+- Make port configurable.
+
 ### 1.2.0
 
 - Show thing prefabHash and prefabName.