+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix crashes when dealing with server name and password.
 - Fix kick/ban being GET requests, should be POST.
 - Add player ping / score / playtime to player payload.
+- Add GET bans and DELETE ban
 
 ### 1.2.0